requirements.txt
absl-py==0.10.0
aio-pika==6.7.1
aiofiles==0.6.0
aiohttp==3.6.3
aiormq==3.3.1
alabaster==0.7.12
appdirs==1.4.4
APScheduler==3.6.3
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
Automat==20.2.0
Babel==2.9.0
beautifulsoup4==4.9.3
bidict==0.21.2
black==20.8b1
bleach==3.3.0
blis==0.7.4
boto3==1.16.42
botocore==1.19.42
cachetools==4.2.0
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.4
characteristic==14.3.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorclass==2.2.0
coloredlogs==14.3
colorhash==1.0.3
ConfigArgParse==1.4
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.7.1
dm-tree==0.1.5
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
fakeredis==1.4.5
fbmessenger==6.0.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
gevent==21.1.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.16
greenlet==1.0.0
grpcio==1.34.0
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.18.1
httptools==0.1.1
httpx==0.15.4
humanfriendly==9.1
hyperlink==21.0.0
idna==2.10
imagesize==1.2.0
incremental==21.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.4.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kafka-python==2.0.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
klein==20.6.0
locationtagger==0.0.1
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mattermostwrapper==2.2
mistune==0.8.4
multidict==4.7.6
murmurhash==1.0.5
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nbsphinx==0.8.2
nest-asyncio==1.5.1
networkx==2.5
newspaper3k==0.2.8
nltk==3.5
numpy==1.16.6
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pamqp==2.3.0
pandoc==1.0.2
pandocfilters==1.4.3
pathlib==1.0.1
pathspec==0.8.1
pathy==0.4.0
Pillow==8.0.1
ply==3.11
pony==0.7.14
preshed==3.0.5
prompt-toolkit==2.0.10
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pydantic==1.7.3
pydot==1.4.1
Pygments==2.8.1
PyJWT==2.0.1
pykwalify==1.7.0
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyTelegramBotAPI==3.7.6
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.5.1
python-telegram-bot==12.8
pytz==2020.4
PyYAML==5.3.1
pyzmq==22.0.3
questionary==1.5.2
rasa==2.3.1
rasa-nlu==0.11.5
rasa-sdk==2.3.1
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.9.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.4.post1
scikit-learn==0.23.2
scipy==1.5.4
sentry-sdk==0.19.5
sgmllib3k==1.0.0
simplejson==3.17.2
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smart-open==3.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
soupsieve==2.2.1
spacy==3.0.5
spacy-legacy==3.0.1
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.22
srsly==2.4.0
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
tensorflow-text==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
testpath==0.4.4
thinc==8.0.2
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tornado==6.1
tqdm==4.50.2
traitlets==5.0.5
Tubes==0.2.0
twilio==6.45.4
Twisted==21.2.0
typed-ast==1.4.2
typeguard==2.10.0
typer==0.3.2
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==3.2.0
urllib3==1.26.2
uvloop==0.14.0
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
zope.event==4.5.0
zope.interface==5.2.0
